[move] #6 TheaterSelectionModalFooter.kt 생성 후 이곳으로 TheaterSelectionModalFooter, ShowSelectedChipsBox, Chip 컴포저블 함수들과 Modifier.dropShadow 함수 이동
